#36 | brief comment

+++ src/components/Life.hpp
@@ -1,12 +1,17 @@
+/**
+ * components/life.hpp
+ * Basic Life component. This should be displayed as an integer, but the internal representation
+ * is a double.
+ */
 #ifndef COMPONENT_LIFE_HPP
 #define COMPONENT_LIFE_HPP
 
 #include "entityx/entityx.h"
 
 struct Life : entityx::Component<Life> {
-  Life(int l = 0) : life(l) {}
+  Life(double l = 0.0) : life(l) {}
 
-  int life;
+  double life;
 };
 
 #endif
